feat(serial-codec): add helpers, tooling and codec claude notes
Add remaining serial, codec and helper assets around the new core:\n\n- Bring in serial helper sources and PlatformIO helper scripts under\n  firmware/v2/src/serial and firmware/v2/pio.\n- Add include/ scaffolding and kid-adding utilities used for\n  generated effect IDs and presets.\n- Commit codec CLAUDE notes and palette docs capturing the HTTP/WS\n  codec behaviour under the EffectId model.\n\nThis completes the firmware-side scaffolding and documentation around\nserial and codec functionality.

+++ firmware/v2/src/codec/WsPluginsCodec.cpp
@@ -131,7 +131,7 @@ void WsPluginsCodec::encodePluginsList(const ::lightwaveos::plugins::PluginManag
     
     // List registered effect IDs
     JsonArray effects = data["effects"].to<JsonArray>();
-    for (uint8_t i = 0; i < plugins::PluginConfig::MAX_EFFECTS; i++) {
+    for (uint16_t i = 0; i < plugins::PluginConfig::MAX_EFFECTS; i++) {
         if (manager.isEffectRegistered(i)) {
             effects.add(i);
         }
